pyproject.toml
[project]
name = "rag-pipeline-airflow"
version = "0.1.0"
description = ""
authors = [
    {name = "shushil girish",email = "girish.sh@northeastern.edu"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "streamlit (>=1.43.2,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic (>=2.10.6,<3.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "litellm (>=1.63.12,<2.0.0)",
    "pinecone (>=6.0.2,<7.0.0)",
    "serpapi (>=0.1.5,<0.2.0)",
    "google-search-results (>=2.4.2,<3.0.0)",
    "langgraph (>=0.3.20,<0.4.0)",
    "langchain (>=0.3.21,<0.4.0)",
    "langchain-core (>=0.3.48,<0.4.0)",
    "openai (>=1.68.2,<2.0.0)",
    "langchain-community (>=0.3.20,<0.4.0)",
    "snowflake-connector-python (>=3.14.0,<4.0.0)",
    "graphviz (>=0.20.3,<0.21.0)",
    "langchain-openai (>=0.3.10,<0.4.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "boto3 (>=1.37.21,<2.0.0)",
    "pyarrow (<19.0.0)",
    "langchain-anthropic (>=0.3.10,<0.4.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "snowflake-sqlalchemy (>=1.7.3,<2.0.0)",
    "langchain-groq (>=0.3.1,<0.4.0)",
    "mistralai (>=1.6.0,<2.0.0)",
    "fpdf (>=1.7.2,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "tavily-python (>=0.5.3,<0.6.0)",
    "google-cloud-aiplatform (>=1.87.0,<2.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "langchain-google-genai (==2.0.0)",
    "xai-grok-sdk (>=0.0.12,<0.0.13)",
    "streamlit-folium (>=0.24.0,<0.25.0)",
    "folium (>=0.19.5,<0.20.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
